apt: move all message output to json-hook
This means that all messages including esm update messages will only
show up when using the apt binary (not apt-get).

This also simplifies some of our code because granularity of the
json-hook config and method names replaces all of the command line
parsing that was previously done to make sure that messages were
printed for the right commands.
